antes de bonus con archivos basura

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: jorgfern <jorgfern@student.42malaga.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/25 18:05:45 by jorgfern          #+#    #+#             */
-/*   Updated: 2023/04/25 18:05:45 by jorgfern         ###   ########.fr       */
+/*   Updated: 2023/04/28 17:57:48 by jorgfern         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i])
